+++ lib/stake.ts
@@ -2,7 +2,7 @@ import { SessionAccountInterface } from '@argent/tma-wallet';
 import stakeAbi from '@/public/stake_abi.json';
 import strkAbi from '@/public/strk_abi.json';
 import { Contract } from 'starknet';
-import { sendTransaction } from '@/lib/transaction';
+import { sendTransaction } from '@/lib/starknet';
 
 const delegationPoolAddress = '0x07134aad6969880f11b2d50e57c6e8d38ceef3a6b02bd9ea44837bd257023f6b';
 const starkAddress = '0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d';
